requirements.txt
discord.py>=2.3.0
python-telegram-bot>=20.0
httpx[socks]>=0.24.0
python-dotenv>=1.0.0
